test: move default config to top level conftest.py
