+++ tests/conftest.py
@@ -1,6 +1,9 @@
+import os
+
 import pytest
 
-from commitizen import cmd
+from commitizen import cmd, defaults
+from commitizen.config import BaseConfig
 
 
 @pytest.fixture(scope="function")
@@ -18,3 +21,15 @@ def tmp_commitizen_project(tmp_git_project):
         tmp_commitizen_cfg_file.write("[tool.commitizen]\n" 'version="0.1.0"\n')
 
         yield tmp_git_project
+
+
+@pytest.fixture()
+def config():
+    _config = BaseConfig()
+    _config.settings.update({"name": defaults.DEFAULT_SETTINGS["name"]})
+    return _config
+
+
+@pytest.fixture()
+def config_path() -> str:
+    return os.path.join(os.getcwd(), "pyproject.toml")
