+++ tests/test_cz_base.py
@@ -1,15 +1,6 @@
 import pytest
 
-from commitizen import defaults
-from commitizen.config import BaseConfig
 from commitizen.cz.base import BaseCommitizen
-
-
-@pytest.fixture()
-def config():
-    _config = BaseConfig()
-    _config.settings.update({"name": defaults.DEFAULT_SETTINGS["name"]})
-    return _config
 
 
 class DummyCz(BaseCommitizen):
