+++ tests/test_cz_jira.py
@@ -1,15 +1,4 @@
-import pytest
-
-from commitizen import defaults
-from commitizen.config import BaseConfig
 from commitizen.cz.jira import JiraSmartCz
-
-
-@pytest.fixture()
-def config():
-    _config = BaseConfig()
-    _config.settings.update({"name": defaults.DEFAULT_SETTINGS["name"]})
-    return _config
 
 
 def test_questions(config):
